sem: Resolve linux compilation issues

+++ semestral/src/data-exporters/csv.cpp
@@ -8,8 +8,8 @@ void CSVDataExporter::exportData() {
   AbstractDataExporter::exportData();
   ofstream file(this->dstPath);
   vector<string> header = this->dataSource->getHeaderVector();
-  int headerSize = this->dataSource->getHeaderSize();
-  for (int i = 0; i < headerSize; i++) {
+  size_t headerSize = this->dataSource->getHeaderSize();
+  for (size_t i = 0; i < headerSize; i++) {
     file << header[i];
     if (i < headerSize - 1) {
       file << ",";
@@ -18,8 +18,8 @@ void CSVDataExporter::exportData() {
   file << endl;
   while (this->dataSource->hasNextRow()) {
     vector<string> row = this->dataSource->getNextRow();
-    int rowSize = row.size();
-    for (int i = 0; i < rowSize; i++) {
+    size_t rowSize = row.size();
+    for (size_t i = 0; i < rowSize; i++) {
       file << row[i];
       if (i < rowSize - 1) {
         file << ",";
