+++ semestral/src/data-sources/abstract.h
@@ -5,6 +5,7 @@
 #include <vector>
 #include <fstream>
 #include <unordered_map>
+#include <memory>
 
 /**
  * Everything that produces relational data should inherit from this class.
@@ -19,7 +20,7 @@ public:
    */
   virtual std::vector<std::string> getHeaderVector() const = 0;
 
-  virtual std::unordered_map<std::string, int> getHeaderMap() const = 0;
+  virtual std::unordered_map<std::string, size_t> getHeaderMap() const = 0;
 
   /**
    *
@@ -27,20 +28,20 @@ public:
    * @throws std::runtime_error if there is no column with given name.
    * @return index of column with given name.
    */
-  virtual int getHeaderIndex(const std::string &name) const = 0;
+  virtual size_t getHeaderIndex(const std::string &name) const = 0;
 
   /**
    *
    * @param index
    * @return name of column with given index.
    */
-  virtual const std::string& getHeaderName(int index) const = 0;
+  virtual const std::string& getHeaderName(size_t index) const = 0;
 
   /**
    *
    * @return number of columns in this relation
    */
-  virtual int getHeaderSize() const = 0;
+  virtual size_t getHeaderSize() const = 0;
 
   /**
    *
@@ -68,10 +69,10 @@ class FileDataSource : public AbstractDataSource {
 public:
   FileDataSource(const std::string &path);
   virtual std::vector<std::string> getHeaderVector() const override;
-  virtual std::unordered_map<std::string, int> getHeaderMap() const override;
-  virtual int getHeaderIndex(const std::string &name) const override;
-  virtual const std::string& getHeaderName(int index) const override;
-  virtual int getHeaderSize() const override;
+  virtual std::unordered_map<std::string, size_t> getHeaderMap() const override;
+  virtual size_t getHeaderIndex(const std::string &name) const override;
+  virtual const std::string& getHeaderName(size_t index) const override;
+  virtual size_t getHeaderSize() const override;
   virtual ~FileDataSource();
   void reset() override;
 protected:
@@ -79,7 +80,7 @@ protected:
   /**
    * Each data source must have a header describing what each column means. This header must be the same size as each row.
    */
-  std::unordered_map<std::string, int> header;
+  std::unordered_map<std::string, size_t> header;
 };
 
 #endif // SEMESTRAL_RESULTSTREAM_H
