+++ semestral/src/data-sources/expressions/projection-expression.cpp
@@ -11,7 +11,7 @@ ProjectionExpression::ProjectionExpression(shared_ptr<AbstractExpression> expres
   for (const std::string &column: columns) {
     indexes[expression->getHeaderIndex(getWrappedColumnName(column))]++;
   }
-  for (int i = 0; i < indexes.size(); i++) {
+  for (size_t i = 0; i < indexes.size(); i++) {
     if (indexes[i] > 1) {
       throw runtime_error("Column " + expression->getHeaderName(i) + " is duplicated");
     }
@@ -22,7 +22,7 @@ ProjectionExpression::ProjectionExpression(shared_ptr<AbstractExpression> expres
 string ProjectionExpression::toSQL() const {
   string sql = "SELECT ";
   vector<string> header = this->getHeaderVector();
-  for (int i = 0; i < header.size(); i++) {
+  for (size_t i = 0; i < header.size(); i++) {
     if (i > 0) {
       sql += ", ";
     }
@@ -40,15 +40,15 @@ vector<string> ProjectionExpression::getHeaderVector() const {
   return indexMapToVector(this->headerMap);
 }
 
-unordered_map<string, int> ProjectionExpression::getHeaderMap() const {
+unordered_map<string, size_t> ProjectionExpression::getHeaderMap() const {
   return this->headerMap;
 }
 
-int ProjectionExpression::getHeaderIndex(const string &name) const {
+size_t ProjectionExpression::getHeaderIndex(const string &name) const {
   return this->headerMap.at(name);
 }
 
-const string &ProjectionExpression::getHeaderName(int index) const {
+const string &ProjectionExpression::getHeaderName(size_t index) const {
   for (auto &pair: this->headerMap) {
     if (pair.second == index) {
       return pair.first;
@@ -57,7 +57,7 @@ const string &ProjectionExpression::getHeaderName(int index) const {
   throw runtime_error("Index not found");
 }
 
-int ProjectionExpression::getHeaderSize() const {
+size_t ProjectionExpression::getHeaderSize() const {
   return this->headerMap.size();
 }
 
@@ -65,7 +65,7 @@ const vector<string> ProjectionExpression::getNextRow() {
   vector<string> header = this->getHeaderVector();
   vector<string> row(header.size());
   vector<string> nextRow = this->expression->getNextRow();
-  for (int i = 0; i < header.size(); i++) {
+  for (size_t i = 0; i < header.size(); i++) {
     row[i] = nextRow[this->expression->getHeaderIndex(getWrappedColumnName(header[i]))];
   }
   return row;
